requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-markdownx==2.0.28
django-pagedown==1.0.6
djangorestframework==3.9.2
Faker==1.0.6
idna==2.8
Markdown==3.1
martor==1.4.0
Pillow==5.4.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
text-unidecode==1.2
urllib3==1.24.3
